requirements.txt
selene=2.0.0rc4
pytest~=8.0.0rc1
requests=2.31.0
allure-pytest=2.13.2
requests==2.31.0
jsonschema==4.20.0
selene==2.0.0rc4
selenium==4.16.0
python-dotenv==1.0.0